This hopefully completes the initial work to add flexibility to the control of link groups.
There are 3 new config variables for each <Group> tag in hapconfig. 

A) hidehomepage - if set to true, hides the group from the homepage completely, 
B) hidehomepagelink - if set to true, hides the navigation link on the homepage. (can be used to do internal links between groups)
C) hidetopmenu - if set to true, hides the group from the dropdown menu on pages e.g. on the booking page