Allow multiple Internal IP Ranges to be specified.
Add functionality to <link> to allow 'access' attribute to be defined - expected values being both, internal or external; indicating whether to display a given tile to internal users, external users or both. Default is Both. This could be used for instance to hide the "my files" tile onsite, when it would be expected end users access their documents via a network drive.