Fix: javascript error if no livetiles data exists.
If you are viewing  HAP externally, and a tile gorup only has internal only tiles, i.e. is an empty group, this code would generate a javascript error due to data[0] being null. Therefore, we only process tile initialisation if there is some tile data.

Nick: you may want to see if you can think of a better way of doing this -> if you have a tile (#foo, viewable in/external) linking to a tile group which has internal only tiles. When viewing externally, clicking on the link takes you to an incorrect tab due to the missing group. Or you may consider that's a case of end user needs to ensure that the tile group link has internal/external properties set appropriately.